♻️ Split board models into domain package
